% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compswap.R
\name{compswap}
\alias{compswap}
\title{Compare swaprinc Models}
\usage{
compswap(data, formula, engine = "stats", .pca_varlist, .n_pca_list, ...)
}
\arguments{
\item{data}{A dataframe}

\item{formula}{A quoted model formula}

\item{engine}{The engine for fitting the model.  Options are "stats" or"lme4".}

\item{.pca_varlist}{A list of pca_vars (see swaprinc documentation)}

\item{.n_pca_list}{A list of n_pca_components (see swaprinc documentation)}

\item{...}{Pass additional arguments to the swaprinc function}
}
\value{
A list containing a list of fitted models and a comparison metrics
data frame.
}
\description{
Compare swaprinc Models
}
\examples{
# Load the iris dataset
data(iris)

# Define the formula
formula <- "Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width"

# Define the pca_varlist
pca_varlist <- list(c("Sepal.Width", "Petal.Length"),
                   c("Sepal.Width", "Petal.Width"))

# Define the n_pca_list
n_pca_list <- list(2, 2)

# Run compswap
compswap_results <- compswap(data = iris,
                            formula = formula,
                            engine = "stats",
                            .pca_varlist = pca_varlist,
                            .n_pca_list = n_pca_list)
}
