% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaprinc.R
\name{swaprinc}
\alias{swaprinc}
\title{Swap in Principal Components}
\usage{
swaprinc(
  data,
  formula,
  engine = "stats",
  pca_vars,
  n_pca_components,
  norun_raw = FALSE,
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{formula}{A quoted model formula}

\item{engine}{The engine for fitting the model.  Options are "stats" or"lme4".}

\item{pca_vars}{Variables to include in the principal component analysis.
These variables will be swapped out for principal components}

\item{n_pca_components}{The number of principal components to include in the
model}

\item{norun_raw}{Include regression on raw variables if TRUE, exclude if FALSE.}

\item{...}{Pass additional arguments}
}
\value{
A list with fitted models
}
\description{
Swap in Principal Components
}
\examples{
data(iris)
res <- swaprinc(iris,
"Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width",
pca_vars = c("Sepal.Width", "Petal.Length", "Petal.Width"),
n_pca_components = 2)
}
