% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaprinc.R
\name{swaprinc_loud}
\alias{swaprinc_loud}
\title{Swap in Principal Components (Loud Version of swaprinc)}
\usage{
swaprinc_loud(
  data,
  formula,
  engine,
  prc_eng,
  pca_vars,
  n_pca_components,
  norun_raw,
  lpca_center,
  lpca_scale,
  lpca_undo,
  gifi_transform,
  gifi_trans_vars,
  gifi_trans_dims,
  no_tresp,
  miss_handler,
  model_options,
  prcomp_options,
  gifi_princals_options,
  gifi_trans_options
)
}
\arguments{
\item{data}{(see swaprinc documentation)}

\item{formula}{(see swaprinc documentation)}

\item{engine}{(see swaprinc documentation)}

\item{prc_eng}{(see swaprinc documentation)}

\item{pca_vars}{(see swaprinc documentation)}

\item{n_pca_components}{(see swaprinc documentation)}

\item{norun_raw}{(see swaprinc documentation)}

\item{lpca_center}{(see swaprinc documentation)}

\item{lpca_scale}{(see swaprinc documentation)}

\item{lpca_undo}{(see swaprinc documentation)}

\item{gifi_transform}{(see swaprinc documentation)}

\item{gifi_trans_vars}{(see swaprinc documentation)}

\item{gifi_trans_dims}{(see swaprinc documentation)}

\item{no_tresp}{(see swaprinc documentation)}

\item{miss_handler}{(see swaprinc documentation)}

\item{model_options}{(see swaprinc documentation)}

\item{prcomp_options}{(see swaprinc documentation)}

\item{gifi_princals_options}{(see swaprinc documentation)}

\item{gifi_trans_options}{(see swaprinc documentation)}
}
\value{
(see swaprinc documentation)
}
\description{
Swap in Principal Components (Loud Version of swaprinc)
}
\examples{
\dontrun{
(see swaprinc documentation)
}
}
\keyword{internal}
