% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaprinc.R
\name{swaprinc}
\alias{swaprinc}
\title{Swap in Principal Components}
\usage{
swaprinc(
  data,
  formula,
  engine = "stats",
  prc_eng = "stats",
  pca_vars,
  n_pca_components,
  norun_raw = FALSE,
  center = TRUE,
  scale. = FALSE,
  lpca_center = "none",
  lpca_scale = "none",
  lpca_undo = FALSE,
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{formula}{A quoted model formula}

\item{engine}{The engine for fitting the model.  Options are 'stats' or 'lme4'.}

\item{prc_eng}{Then engine or extracting principal components.  Options are
'stats' and 'Gifi'.}

\item{pca_vars}{Variables to include in the principal component analysis.
These variables will be swapped out for principal components}

\item{n_pca_components}{The number of principal components to include in the
model}

\item{norun_raw}{Include regression on raw variables if TRUE, exclude if FALSE.}

\item{center}{Set center parameter for prcomp}

\item{scale.}{Set scale. parameter for prcomp}

\item{lpca_center}{Center data as in the LearnPCA Step-by-Step PCA vignette.  Only
numeric variables will be included in the centering.  Parameter takes values
'all' to center raw and pca variables, 'raw' to only center variables for the
raw variable model fitting, 'pca' to only center pca_vars before pca regression
model fitting, and 'none' to skip lpca centering.}

\item{lpca_scale}{Scale data as in the LearnPCA Step-by-Step PCA vignette.  Only
numeric variables will be included in the scaling.  Parameter takes values
'all' to scale raw and pca variables, 'raw' to only scale variables for the
raw variable model fitting, 'pca' to only scale pca_vars before pca regression
model fitting, and 'none' to skip lpca scaling.}

\item{lpca_undo}{Undo centering and scaling of pca_vars as in the LearnPCA
Step-by-Step PCA vignette.}

\item{...}{Pass additional arguments}
}
\value{
A list with fitted models
}
\description{
Swap in Principal Components
}
\examples{
data(iris)
res <- swaprinc(iris,
"Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width",
pca_vars = c("Sepal.Width", "Petal.Length", "Petal.Width"),
n_pca_components = 2)
}
